+++ src/main/java/br/com/onesystem/domain/NotaRecebida.java
@@ -41,7 +41,7 @@ public class NotaRecebida extends Nota implements Serializable {
             Cotacao cotacao, List<ValorPorCotacao> valorPorCotacao,
             BigDecimal desconto, BigDecimal acrescimo, BigDecimal despesaCobranca,
             BigDecimal frete, BigDecimal aFaturar, BigDecimal totalEmDinheiro, Nota notaDeOrigem, Date emissao,
-            Caixa caixa, Usuario usuario, FaturaRecebida faturaRecebida, ConhecimentoDeFrete conhecimentoDeFrete, Filial filial, PedidoAFornecedores pedidoAFornecedores, Integer numeroNF) throws DadoInvalidoException {
+            Caixa caixa, Usuario usuario, FaturaRecebida faturaRecebida, ConhecimentoDeFrete conhecimentoDeFrete, Filial filial, PedidoAFornecedores pedidoAFornecedores, Long numeroNF) throws DadoInvalidoException {
         super(id, pessoa, operacao, itens, formaDeRecebimento, listaDePreco, cobrancas, cotacao, valorPorCotacao, desconto, acrescimo, despesaCobranca, frete, aFaturar, totalEmDinheiro, notaDeOrigem, emissao, caixa, usuario, filial, numeroNF);
         this.faturaRecebida = faturaRecebida;
         this.conhecimentoDeFrete = conhecimentoDeFrete;
